Cart Contents Saved and Displayed
The Cart class has been created as an object to store and process data related to the customer's cart. The cart contains a dictionary storing a product and the associated quantity of that item.

The cart also features functions to add to the dictionary, clear the dictionary, get the subtotal of all items in the cart, and convert the cart contents to a format (List<List<string>>) that can be used in a table.

A new table is also created when the contents of a cart is greater than 0. Currently it detects the subtotal, which will have a flaw if the price of an item is $0. This could easily be corrected, but this fix will not be applied here.