Cart Totals
The cart can now calculate discounts and return the value of each discount to be displayed in a table, as well as the complete cart price.

As there is only 1 discount, the GetValidDiscounts function only return the BOGO50% red flower discount.

The cart can also now return the shipping price for the cart judging by the subtotal, and factors in discounts before calculating shipping.

These functions are then used in a new function that returns a table formated list of prices for the subtotal, discounts (conditional), shipping price, and complete total.