ReadMe and Comments Update + ProductTable
The ReadMe and comments in all files have been fleshed out and updated to go into depth about how the code works and follows principles of OOP code.

Additionally, to help demonstrate inheritance and polymorphism and the relations between 2 classes, the Table class is now a "generic" Table class, and a new ProductTable class has been added that inherits from the Table class. This class takes the CreateTableStringFromProducts method and moves it into this specialized Table type.